fix(providers): remove unnecessary u64 cast in retry jitter calc
Clippy flagged `as u64` on an expression that already returns u64.

+++ crates/kestrel-providers/src/retry.rs
@@ -482,7 +482,7 @@ pub fn backoff_with_jitter(base: Duration, attempt: u32, max_delay: Duration) ->
     if jitter_max == 0 {
         return raw;
     }
-    let jitter = (jitter_frac % jitter_max) as u64;
+    let jitter = jitter_frac % jitter_max;
     let total = (raw.as_millis() as u64 + jitter).min(max_delay.as_millis() as u64);
     Duration::from_millis(total)
 }
